Add a test for baremetal.
This modifies the virt-what script slightly so that a test
root can be passed, which causes all tests to be run relative
to that root directory (similar to a chroot).

We then provide a root directory with just enough files in it
to convince virt-what that it is a baremetal machine.

Note that the data in the files is completely real.  We do
not intend to add a test unless we can get data from the
real system.