Add IA64-specific tests for virtualization (RHBZ#707508).
Xen IA64 HVM does not provide a simple way to detect the
hypervisor.

If we are lucky and PV-on-HVM drivers are installed and loaded,
then there will be a /sys/bus/xen directory.

Otherwise we add a timing attack to see if we can detect generic
virtualization.  Since this is a big hack, it is only enabled on IA64
and only used when all other methods of detection have failed.

This commit also adds `uname -p` to all tests (enabling arch detection
so we only run the above tests on IA64).