Direct output from fail() to stderr
This allows callers to more easily differentiate valid output from
errors.  A recent example is facter.  When calling virt-what for its
virtual fact, it gets 'virt-what: this script must be run as root' as
the fact value when run as non-root:

$ facter | grep ^virtual
virtual => virt-what: this script must be run as root

This is mildly related to RHBZ #719611¹.

¹ https://bugzilla.redhat.com/show_bug.cgi?id=719611